pyproject.toml
[project]
name = "steam_hook"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "python-dotenv>=1.1.1",
    "steam[client]>=1.4.4",
]
